\name{baseball.attendance}
\alias{baseball.attendance}
\docType{data}
\title{Attendance at baseball teams.}
\description{
Average attendance for home and away games for all major league
baseball games in the first half of the 2010 season.
}
\usage{
baseball.attendance
}
\format{
  A data frame with 30 observations on the following 8 variables.
  \describe{
  \item{Team}{name of team}
  \item{League}{league where the team belongs}
  \item{N.Home}{number of home games}
  \item{Avg.Home}{average attendance for home games}
  \item{Pct.Home}{percentage of park capacity for home games}
  \item{N.Away}{number of away games}
  \item{Avg.Away}{average attendance for away games}
  \item{Pct.Away}{percentage of park capacity for away games}
}
}
\source{www.espn.com web site}
\keyword{datasets}
