\name{batting.history}
\alias{batting.history}
\docType{data}
\title{Yearly Batting Statistics for Major League Baseball}
\description{
Year to year batting statistics for all years of professional American baseball.
}
\usage{
batting.history
}
\format{
  A data frame with 140 observations on the following 27 variables.
  \describe{
  \item{Year}{baseball season year}
  \item{Tms}{number of teams}
  \item{N.Bat}{number of players used}
  \item{BatAge}{average age of batters}
  \item{R}{runs scored per game}
  \item{G}{number of games played}
  \item{PA}{number of plate appearances per game}
  \item{AB}{number of at-bats per game}
  \item{H}{average number of hits per game }
  \item{X2B}{average number of doubles per game}
  \item{X3B}{average number of triples per game}
  \item{HR}{average number of home runs per game}
  \item{RBI}{average number of runs-batted-in per game}
  \item{SB}{average number of stolen bases per game}
  \item{CS}{average number of caught stealing per game}
  \item{BB}{average number of base-on-balls per game}
  \item{SO}{average number of strikeouts per game}
  \item{BA}{average batting average}
  \item{OBP}{average on-base percentage}
  \item{SLG}{average slugging percentage}
  \item{OPS}{average OBP + SLG}
  \item{TB}{average total bases per game}
  \item{GDP}{average grounded into double plays per game}
  \item{HBP}{average hit by pitch per game}
  \item{SH}{average sacrifice hits per game}
  \item{SF}{average sacrifice fliew per game}
  \item{IBB}{average number of intentional walks per game}
}
}
\source{www.baseball-reference.com web site}
\keyword{datasets}
