\name{beatles}
\alias{beatles}
\docType{data}
\title{Lengths of songs from Beatles' albums.}
\description{
Length of songs from a selection of albums from the popular
pop and rock group The Beatles.
}
\usage{
beatles
}
\format{
  A data frame with 113 observations on the following 2 variables.
  \describe{
  \item{time}{time in seconds of the song}
  \item{album}{name of Beatles' album}
  }
}
\source{Collected by the author from the album descriptions}
\keyword{datasets}
