\name{boston.marathon}
\alias{boston.marathon}
\docType{data}
\title{Boston Marathon completion times of women of different ages.}
\description{
Completion times for a sample of women at the 2001 Boston Marathon.
}
\usage{
boston.marathon
}
\format{
  A data frame with 108 observations on the following 2 variables.
  \describe{
  \item{time}{completion time in minutes}
  \item{age}{age of the participant}
  }
}
\source{www.bostonmarathon.org website}
\keyword{datasets}
