\name{boston.marathon.wtimes}
\alias{boston.marathon.wtimes}
\docType{data}
\title{Boston marathon winning times}
\description{
Winning time in mens Boston marathon through the years.
}
\usage{
boston.marathon.wtimes
}
\format{
  A data frame with 99 observations on the following 2 variables.
  \describe{
  \item{year}{year}
  \item{minutes}{winning time in minutes}
  }
}
\source{2001 ESPN Information Please Sports Almanac}
\keyword{datasets}

