\name{braves.attendance}
\alias{braves.attendance}
\docType{data}
\title{Attendance at games of a baseball team}
\description{
Attendance of Atlanta Braves home games during a particular season.
}
\usage{
braves.attendance
}
\format{
  A data frame with 72 observations on the following 2 variables.
  \describe{
  \item{Game}{game number}
  \item{Attendance}{attendance count}
  }
}
\source{www.espn.com website}
\keyword{datasets}

