\name{church.tseries}
\alias{church.tseries}
\docType{data}
\title{Time series of worship attendance}
\description{
Time series of weekly worship attendance from a church in Ohio, U.S.
}
\usage{
church.tseries
}
\format{
  A data frame with 209 observations on the following 3 variables.
  \describe{
  \item{year}{the year}
  \item{month}{the month}
  \item{worship}{count of worship attendance}
 }}
\source{Collected by the author.}
\keyword{datasets}