\name{college.ratings}
\alias{college.ratings}
\docType{data}
\title{Ratings of National Universities in the U.S.}
\description{
Ratings for National Universities based on a 2001 survey.
}
\usage{
college.ratings
}
\format{
  A data frame with 249 observations on the following 12 variables.
  \describe{
  \item{School}{name of university}
  \item{Tier}{tier grouping of the univesity}
  \item{Type}{private or public institution}
  \item{Reputation}{measure of academic reputation}
  \item{F.retention}{average freshmen retention rate}
  \item{a.grad.rate}{percentage of freshmen who graduated within a six-year period}
  \item{Under.20}{percentage of undergraduate classes with fewer than 20 students}
  \item{Over.50}{percentage of undergraduate classes with 50 students or more enrolled}
  \item{Full.time}{percentage of the total number of faculty employed on a full-time basis}
  \item{Top.10}{proportion of students enrolled who graduated in the top 10 percent of their high school class}
  \item{Acc.rate}{ratio of the number of students admitted to the number of applicants}
  \item{Alum.giving}{average percent of undergraduate alumni of record who donated money to the college or university}
}}
\source{U.S. News and World Report}
\keyword{datasets}
