\name{farms}
\alias{farms}
\docType{data}
\title{Number of farms in the states of the U.S.}
\description{
The number of farms (in 1000s) for each of the 50 states 
in the United States in 1999.
}
\usage{
farms
}
\format{
  A data frame with 50 observations on the following 2 variables.
  \describe{
  \item{state}{name of the state}
  \item{count}{count of the number of farms}
  }
}
\source{2001 New York Times Almanac}
\keyword{datasets}
