\name{football}
\alias{football}
\docType{data}
\title{American football scores}
\description{
Scores for a group of American football games.}

\usage{
football
}
\format{
  A data frame with 465 observations on the following 2 variables.
  \describe{
  \item{winner}{score of winning team}
  \item{loser}{score of losing team}
   }}
\source{Collected by the author}
\keyword{datasets}
