\name{gestation.periods}
\alias{gestation.periods}
\docType{data}
\title{Gestation periods for different animals.}
\description{
Average gestation periods for different animals.
}
\usage{
gestation.periods
}
\format{
  A data frame with 43 observations on the following 2 variables.
  \describe{
  \item{Animal}{name of animal}
  \item{Period}{average gestation period}
  }
}
\source{World Almanac and Book of Facts 2001}
\keyword{datasets}
