\name{grandma.19.40}
\alias{grandma.19.40}
\docType{data}
\title{Grandmas Marathon completion times}
\description{
Completion times in running Grandmas Marathon (Duluth, Minnesota) 
for women of ages 19 to 34}

\usage{
grandma.19.40
}
\format{
  A data frame with 1208 observations on the following 1 variable.
  \describe{
  \item{time}{completion time in seconds}
    }
}
\source{www.grandmamarathon.org web site}
\keyword{datasets}
