\name{heaviest.fish}
\alias{heaviest.fish}
\docType{data}
\title{Fish world record catches}
\description{
Weights in pounds of world record catches of saltwater fish.
}
\usage{
heaviest.fish
}
\format{
  A data frame with 25 observations on the following 2 variables.
  \describe{
  \item{Species}{name of country}
  \item{Weight}{region of the world}
  }
}
\source{
World Almanac and Book of Facts 2008
}
\keyword{datasets}
