\name{home.prices}
\alias{home.prices}
\docType{data}
\title{Home sales prices in the U.S.}
\description{
Average sales prices of houses for 21 cities in the U.S. 
}
\usage{
home.prices
}
\format{
  A data frame with 21 observations on the following 3 variables.
  \describe{
  \item{City}{name of the city}
  \item{y1985}{average home price in 1985}
  \item{y2000}{average home price in 2000}
  }
 }
\source{2001 New York Times Almanac}
\keyword{datasets}
