\name{homeruns.2000}
\alias{homeruns.2000}
\docType{data}
\title{Team home run numbers for different seasons.}
\description{
Total number of home runs hit all teams in the major league baseball
league for the years 1900,1910,..., 2000.
}
\usage{
homeruns.2000
}
\format{
  A data frame with 210 observations on the following 2 variables.
  \describe{
  \item{YEARS}{season}
  \item{HOMERUNS}{count of team home runs}
  }
}
\source{Sean Lahman's database at www.baseball1.com}
\keyword{datasets}
