\name{homeruns.61}
\alias{homeruns.61}
\docType{data}
\title{Home run counts in 1961}
\description{
The number of the number of home runs hit by all major league players 
in the year 1961.
}
\usage{
homeruns.61
}
\format{
  A data frame with 698 observations on the following 1 variables.
  \describe{
  \item{HR}{count of home runs}
  }
}
\source{Sean Lahman's database at www.baseball1.com}
\keyword{datasets}
