\name{immigrants}
\alias{immigrants}
\docType{data}
\title{Immigrant counts to US.}
\description{
Immigrant counts to US from various countries.
}
\usage{
immigrants
}
\format{
  A data frame with 41 observations on the following 4 variables.
  \describe{
  \item{Country}{name of country}
  \item{Region}{region of the world}
  \item{Count.1998}{the count of immigrants in 1998}
 \item{Count.2008}{the count of immigrants in 2008}
  }
}
\source{
2010 New York Times Almanac
}
\keyword{datasets}
