\name{island.areas}
\alias{island.areas}
\docType{data}
\title{Areas of islands from different continents.}
\description{
Size, name, and surrounding ocean for 59 different islands.
}
\usage{
islands.areas
}
\format{
  A data frame with 59 observations on the following 3 variables.
  \describe{
  \item{Ocean}{ocean where the island lies}
  \item{Name}{name of the island}
  \item{Area}{area of the island in square miles)}
}}
\source{World Almanac and Book of Facts}
\keyword{datasets}
