\name{lake}
\alias{lake}
\docType{data}
\title{Lake measurements}
\description{
Measurements of lakes in the Vilas and Oneida counties 
of northern Wisconsin (from Minitab dataset collection).
}
\usage{
lake
}
\format{
  A data frame with 71 observations on the following 5 variables.
  \describe{
  \item{Area}{area of lake in acres}
  \item{Depth}{maximum depth of lake in feet}
  \item{PH}{pH (acidity) measurement}
  \item{Wshed}{watershed area in square miles}
  \item{Hions}{concentration of hydrogen ions}
}}
\source{Minitab data collection}
\keyword{datasets}

