\name{mortality.rates}
\alias{mortality.rates}
\docType{data}
\title{Infant mortality rates of countries}
\description{
Infant mortality rates in 2008 for 62 countries (Afghanistan through Guyana).
}
\usage{
mortality.rates
}
\format{
  A data frame with 62 observations on the following 2 variables.
  \describe{
  \item{Country}{name of country}
  \item{Rate}{infant mortality rate}
  }
}
\source{2010 New York Times Almanac}
\keyword{datasets}

