\name{olympics.run}
\alias{olympics.run}
\docType{data}
\title{Olympics running times.}
\description{
Winning time in Olympics mens running race classified by length of race and year.
}
\usage{
olympics.run
}
\format{
  A data frame with 9 observations on the following 6 variables.
  \describe{
  \item{year}{year of the Olympics}
  \item{X100m}{winning time in seconds for 100m}
  \item{X200m}{winning time in seconds for 200m}
  \item{X400m}{winning time in seconds for 400m}
  \item{X800m}{winning time in seconds for 800m}
  \item{X1500m}{winning time in seconds for 1500m}
  }}
\source{World Almanac and Book of Facts 2008}
\keyword{datasets}
