\name{olympics.speed.skating}
\alias{olympics.speed.skating}
\docType{data}
\title{Olympics speed skating times.}
\description{
Winning time in Olympics mens speed skating race classified by length of race and year.
}
\usage{
olympics.speed.skating
}
\format{
  A data frame with 9 observations on the following 6 variables.
  \describe{
  \item{YEAR}{year of the Olympics}
  \item{X500m}{winning time in seconds for 500m}
  \item{X1000m}{winning time in seconds for 1000m}
  \item{X1500m}{winning time in seconds for 1500m}
  \item{X5000m}{winning time in seconds fro 5000m}
  \item{X10000m}{winning time in seconds for 10000m}
  }}
\source{World Almanac and Book of Facts 2008}
\keyword{datasets}