\name{olympics.swim}
\alias{olympics.swim}
\docType{data}
\title{Olympics swimming times.}
\description{
Winning time in Olympics womens swimming race classified by length  of race and year.
}
\usage{
olympics.swim
}
\format{
  A data frame with 10 observations on the following 5 variables.
  \describe{
  \item{YEAR}{year of the Olympics}
  \item{X100m}{winning time in seconds for 100m}
  \item{X200m}{winning time in seconds for 200m}
  \item{X400m}{winning time in seconds for 400m}
  \item{X800m}{winning time in seconds for 800m}
  }
}
\source{2001 ESPN Information Please Sports Almanac}
\keyword{datasets}
