\name{pitching.history}
\alias{pitching.history}
\docType{data}
\title{Yearly Pitching Statistics for Major League Baseball}
\description{
Year to year pitching statistics for all years of professional American baseball.
}
\usage{
batting.history
}
\format{
  A data frame with 140 observations on the following 31 variables.
  \describe{
  \item{Year}{baseball season year}
  \item{Tms}{number of teams}
  \item{N.Pitch}{number of pitchers used}
  \item{PitchAge}{average age of pitchers}
  \item{R}{runs scored per game per team}
  \item{ERA}{average earned run average}
  \item{G}{number of games played}
  \item{GF}{proportion of games finished}
  \item{CG}{proportion of complete games}
  \item{SHO}{proportion of shutouts}
  \item{SV}{number of saves per game}
  \item{IP}{innings pitched per game}
  \item{H}{average number of hits}
  \item{ER}{average number of earned runs}
  \item{HR}{average number of home runs}
  \item{BB}{average number of base-on-balls}
  \item{IBB}{average number of intentional walks}
  \item{SO}{average number of strikeouts}
  \item{HBP}{average number of hit by pitch}
  \item{BK}{average number of balks}
  \item{WP}{average number of wild pitchers}
  \item{BF}{average number of batters faced}
  \item{WHIP}{average hits + walks per inning}
  \item{H.9}{average number of hits per 9 innings}
  \item{HR.9}{average number of home runs per 9 innings}
  \item{BB.9}{average number of walks per 9 innings}
  \item{SO.9}{average number of strikeouts per 9 innings}
  \item{SO.BB}{strikeouts to walks ratio}
  \item{E}{average number of errors}
  \item{Attendance}{total attendance}
  \item{Attend.G}{attendance per game}
}
}
\source{www.baseball-reference.com web site}
\keyword{datasets}
