\name{pop.change}
\alias{pop.change}
\docType{data}
\title{Population change for all states in the U.S.}
\description{
Population change for all states in the United States between the
years 2000 and 2009.
}
\usage{
pop.change
}
\format{
  A data frame with 51 observations on the following 5 variables.
  \describe{
  \item{State}{name of state}
  \item{Abbrev}{state abbreviation}
  \item{Pop.2000}{population of state in 2000}
  \item{Pop.2009}{population of state in 2009}
  \item{Pct.change}{percentage change in population}
  }
}
\source{2010 New York Times Almanac}
\keyword{datasets}
