\name{pop.densities}
\alias{pop.densities}
\docType{data}
\title{Population densities of states for different years.}
\description{
Population density of each state in the United States for the years
1960, 1970, 1980, 1990, 2000, and 2008.
}
\usage{
pop.densities
}
\format{
  A data frame with 51 observations on the following 7 variables.
  \describe{
  \item{State}{name of state}
  \item{Abbrev}{state abbreviation}
  \item{y1960}{population density of state in 1960}
  \item{y1970}{population density of state in 1960}
  \item{y1980}{population density of state in 1980}
  \item{y1990}{population density of state in 1990}
  \item{y2000}{population density of state in 2000}
  \item{y2008}{population density of state in 2008}
  }
}
\source{2010 New York Times Almanac}
\keyword{datasets}
