\name{pop.england}
\alias{pop.england}
\docType{data}
\title{England population}
\description{
Population of England and Wales over the years.
}
\usage{
pop.england
}
\format{
  A data frame with 14 observations on the following 2 variables.
  \describe{
  \item{YEAR}{year}
  \item{POPULATION}{population count}
 }}
\source{Tukey, EDA}
\keyword{datasets}
