\name{salaries}
\alias{salaries}
\docType{data}
\title{Salaries of different professions in different cities.}
\description{
1979 salaries of seven different professions in six cities.
}
\usage{
salaries
}
\format{
  A data frame with 42 observations on the following 3 variables.
  \describe{
  \item{Salary}{salary in hundreds of Swiss francs}
  \item{City}{name of city}
  \item{Profession}{name of profession}
}}
\source{unknown}
\keyword{datasets}