\name{snowfall}
\alias{snowfall}
\docType{data}
\title{Snowfall amounts of two cities}
\description{
Seasonal snowfall (in inches) in Buffalo, New York and Cairo, Illinois
from 1918-19 through 1937-38.}
\usage{
snowfall
}
\format{
  A data frame with 40 observations on the following 3 variables.
  \describe{
  \item{Season}{season where 1918 indicates the season 1918-19}
  \item{City}{name of the city}
  \item{Snowfall}{snowfall in inches}
  }
}
\source{Tukey (1977), Exploratory Data Analysis, exhibit 11 of chapter 4}
\keyword{datasets}
