\name{temperatures}
\alias{temperatures}
\docType{data}
\title{Temperatures for different cities.}
\description{
Two-table of average temperatures of cities for different months.}
\usage{
temperatures
}
\format{
  A data frame with 5 observations on the following 5 variables.
  \describe{
  \item{City}{city}
  \item{January}{average temperature in January}
  \item{April}{average temperature in April}
  \item{July}{average temperature in July}
  \item{October}{average temperature in October}
  }
  }
\source{www.weatherbase.com website}
\keyword{datasets}
