\name{tukey.24a}
\alias{tukey.24a}
\docType{data}
\title{Tukey straightening exercise 24a}
\description{
Data from exercise in Tukey, EDA.
}
\usage{
tukey.24a
}
\format{
  A data frame with 16 observations on the following 3 variables.
  \describe{
  \item{latitude}{latitude}
  \item{index}{index}
  \item{temp}{temp}
 }}
\source{Tukey, EDA}
\keyword{datasets}
