\name{tukey.24b}
\alias{tukey.24b}
\docType{data}
\title{Tukey straightening exercise 24b}
\description{
Data from exercise in Tukey, EDA.
}
\usage{
tukey.24b
}
\format{
  A data frame with 34 observations on the following 2 variables.
  \describe{
  \item{diameter}{diameter}
  \item{stretch}{stretch}
 }}
\source{Tukey, EDA}
\keyword{datasets}
