\name{tukey.26a}
\alias{tukey.26a}
\docType{data}
\title{Tukey straightening exercise 24a}
\description{
Measurement of a certain impurity in DDT; 
change of scale factor with temperature.
}
\usage{
tukey.26a
}
\format{
  A data frame with 14 observations on the following 2 variables.
  \describe{
  \item{temp}{temp}
  \item{rate}{rate}
 }}
\source{Tukey, EDA}
\keyword{datasets}
