\name{tukey.26b}
\alias{tukey.26b}
\docType{data}
\title{Tukey straightening exercise 24b}
\description{
Demand deposits in post-office savings accounts in Switzerland.}
\usage{
tukey.26b
}
\format{
  A data frame with 29 observations on the following 2 variables.
  \describe{
  \item{year}{year}
  \item{deposits}{deposits}}
 }
\source{Tukey, EDA}
\keyword{datasets}
