\name{tukey.26c}
\alias{tukey.26c}
\docType{data}
\title{Tukey straightening exercise 24c}
\description{
Revenue passenger miles on U.S. passenger airlines.}
\usage{
tukey.26c
}
\format{
  A data frame with 24 observations on the following 2 variables.
  \describe{
  \item{year}{year}
  \item{miles}{miles}
 }}
\source{Tukey, EDA}
\keyword{datasets}
