\name{us.pop}
\alias{us.pop}
\docType{data}
\title{Population of United States}
\description{
Population of the United States over time.
}
\usage{
us.pop
}
\format{
  A data frame with 22 observations on the following 2 variables.
  \describe{
  \item{YEAR}{year}
  \item{POP}{population count in millions}
  }
 }
\source{Tukey, EDA}
\keyword{datasets}
